ContosoOutdoorStore/PromptFlow/requirements.txt
azure.search.documents
azure.cosmos
promptflow.connections